+++ packages/create-svelte/shared/+playwright+typescript/playwright.config.ts
@@ -4,7 +4,8 @@ const config: PlaywrightTestConfig = {
 	webServer: {
 		command: 'npm run build && npm run preview',
 		port: 4173
-	}
+	},
+	testMatch: 'tests/**/.*(test|spec).(js|ts)'
 };
 
 export default config;