+++ packages/create-svelte/shared/+vitest/vite.config.js
@@ -1,9 +1,11 @@
 import { sveltekit } from '@sveltejs/kit/vite';
-import type { UserConfig } from 'vite';
 
 /** @type {import('vite').UserConfig} */
-const config: UserConfig = {
-	plugins: [sveltekit()]
+const config = {
+	plugins: [sveltekit()],
+	test: {
+		include: ['src/**/*.{test,spec}.{js,ts}']
+	}
 };
 
 export default config;