+++ packages/kit/src/runtime/client/client.js
@@ -1192,8 +1192,12 @@ export function create_client({ target, base }) {
 			}, 20);
 		});
 
-		target.addEventListener('touchstart', (event) =>
-			prepare_and_preload(/** @type {Element} */ (event.composedPath()[0]), 1)
+		target.addEventListener(
+			'touchstart',
+			(event) => {
+				prepare_and_preload(/** @type {Element} */ (event.composedPath()[0]), 1);
+			},
+			{ passive: true }
 		);
 
 		target.addEventListener('keydown', (event) => {
@@ -1775,8 +1779,8 @@ function reset_focus() {
 
 if (__SVELTEKIT_DEV__) {
 	// Nasty hack to silence harmless warnings the user can do nothing about
-	const warn = console.warn;
-	console.warn = (...args) => {
+	const console_warn = console.warn;
+	console.warn = function warn(...args) {
 		if (
 			args.length === 1 &&
 			/<(Layout|Page)(_[\w$]+)?> was created (with unknown|without expected) prop '(data|form)'/.test(
@@ -1785,6 +1789,6 @@ if (__SVELTEKIT_DEV__) {
 		) {
 			return;
 		}
-		warn(...args);
+		console_warn(...args);
 	};
 }